On branch main  Your branch is ahead of 'origin/main' by 4 commits.    (use "git push" to publish your local commits)
 Changes to be committed:
	modified:   src/empresa_agro/Almacen.java

+++ src/empresa_agro/Almacen.java
@@ -40,7 +40,7 @@ public class Almacen {
             do {
 
                 System.out.println("Indique el precio por kilo del producto: ");
-                precioKilos = Main.SC.nextInt();
+                precioKilos = Main.SC.nextDouble();
 
                 if (precioKilos <= 0) {
